==> 18_countoutcomes.Rmd <==
# Modeling a Count Outcome in Ohio SMART

In this chapter, I use a count outcome (# of poor physical health days out of the last 30) to demonstrate regression models for count outcomes. 

Methods discussed in the chapter include the following:

- Ordinary Least Squares
- Poisson Regression
- Overdispersed Quasi-Poisson Regression
- Negative Binomial Regression
- Zero-Inflated Poisson Regression
- Zero-Inflated Negative Binomial Regression
- Hurdle Models with Poisson counts
- Hurdle Models with Negative Binomial counts
- Tobit Regression

## Preliminaries

```{r}
library(boot)
library(lmtest)
library(sandwich)
library(countreg)
library(VGAM)
```

```{r import}
smart_oh <- read.csv("data/smart_ohio.csv") %>% tbl_df
```

## A subset of the Ohio SMART data

Let's consider the following data. I'll include the subset of all observations in `smart_oh` with complete data on these 14 variables.

Variable | Description
--------: | -----------------------------------------------
`SEQNO` | Subject identification code
`MMSANAME` | Name of statistical area
`genhealth` | Five categories (E, VG, G, F, P) on general health
`physhealth` | Now thinking about your physical health, which includes physical illness and injury, for how many days during the past 30 days was your physical health not good?
`menthlth` | Now thinking about your mental health, which includes stress, depression, and problems with emotions, for how many days during the past 30 days was your mental health not good?
`healthplan` | 1 if the subject has any kind of health care coverage, 0 otherwise
`costprob` | 1 indicates Yes to "Was there a time in the past 12 months when you needed to see a doctor but could not because of cost?"
`sleephrs` | average amount of sleep the subject gets in a 24-hour period
`agegroup` | 13 age groups from 18 through 80+
`female` | 1 if subject is female
`incomegroup` | 8 income groups from < 10,000 to 75,000 or more
`bmi` | body-mass index
`smoke100` | 1 if Yes to "Have you smoked at least 100 cigarettes in your entire life?"
`alcdays` | # of days out of the past 30 on which the subject had at least one alcoholic drink

```{r}
sm_oh_A <- smart_oh %>%
    select(SEQNO, MMSANAME, genhealth, physhealth, 
           menthealth, healthplan, costprob, sleephrs, 
           agegroup, female, incomegroup, bmi, smoke100, 
           alcdays) %>%
    drop_na
```

### Is age group associated with `physhealth`?

```{r}
ggplot(sm_oh_A, aes(x = agegroup, y = physhealth)) +
    geom_violin(col = "blue")
```

It's hard to see much of anything here. The main conclusion seems to be that 0 is by far the most common response.

Here's a table by age group of:

- the number of respondents in that age group,
- the group's mean `physhealth` response (remember that these are the number of poor physical health days in the last 30),
- their median `physhealth` response (which turns out to be 0 in each group), and
- the percentage of group members who responded 0.

```{r}
sm_oh_A %>% group_by(agegroup) %>%
    summarize(n = n(), mean = round(mean(physhealth),2), 
              median = median(physhealth),
              percent_0s = round(100*sum(physhealth == 0)/n,1))
```

We can see a real change between the 45-49 age group and the 50-54 age group. The mean difference is clear from the table above, and the plot below (of the percentage with a zero response) in each age group identifies the same story.

```{r}
sm_oh_A %>% group_by(agegroup) %>%
    summarize(n = n(), 
              percent_0s = round(100*sum(physhealth == 0)/n,1)) %>%
    ggplot(aes(y = agegroup, x = percent_0s)) +
    geom_label(aes(label = percent_0s)) +
    labs(x = "% with no Bad Physical Health Days in last 30",
         y = "Age Group")
```

It looks like we have a fairly consistent result in the younger age range (18-49) or the older range (50+). On the theory that most of the people reading this document are in that younger range, we'll focus on those respondents in what follows.

## Exploratory Data Analysis (in the 18-49 group)

We want to predict the 0-30 `physhealth` count variable for the 18-49 year old respondents. 

To start, we'll use two predictors:

- the respondent's body mass index, and
- whether the respondent has smoked 100 cigarettes in their lifetime.

We anticipate that each of these variables will have positive associations with the `physhealth` score. That is, heavier people, and those who have used tobacco will be less healthy, and thus have higher numbers of poor physical health days.

### Build a subset of those ages 18-49

First, we'll identify the subset of respondents who are between 18 and 49 years of age.

```{r} 
sm_oh_A_young.raw <- sm_oh_A %>%
    filter(agegroup %in% c("18-24", "25-29", "30-34", 
                           "35-39", "40-44", "45-49")) %>%
    droplevels()

sm_oh_A_young.raw %>% 
    select(physhealth, bmi, smoke100, agegroup) %>%
    summary
```

Now, let's look more closely at the distribution of these variables, starting with our outcome.

### Distribution of the Outcome

What's the distribution of `physhealth`?

```{r}
ggplot(sm_oh_A_young.raw, aes(x = physhealth)) +
    geom_histogram(binwidth = 1, fill = "red", col = "white")
```

```{r}
sm_oh_A_young.raw %>%
    count(physhealth == 0, physhealth == 30)
```

Most of our respondents said zero, the minimum allowable value, although there is also a much smaller bump at 30, the maximum value we will allow.

Dealing with this distribution is going to be a bit of a challenge. We will develop a series of potential modeling approaches for this sort of data, but before we do that, let's look at the distribution of our other two variables, and the pairwise associations, in a scatterplot matrix.

### Initial Scatterplot Matrix

```{r}
GGally::ggpairs(sm_oh_A_young.raw %>% 
                    select(bmi, smoke100, physhealth))
```

### Dropping the subject with an unreasonably large `bmi`

It looks like there is a very large outlier in the `bmi` data. This subject appears to be a full 40 kg/m^2^ larger than the next largest subject. Without a good explanation for that discrepancy, I will remove that subject before fitting models. I'm also going to center the `bmi` variable to help me interpret the final models later.

```{r}
sm_oh_A_young <- sm_oh_A_young.raw %>%
    mutate(bmi_c = bmi - mean(bmi)) %>%
    filter(bmi < 80)

nrow(sm_oh_A_young)
```

### Revised (Final) Scatterplot Matrix

Now, here's the revised scatterplot matrix for those `r nrow(sm_oh_A_young)` subjects, now using the centered `bmi` data captured in the `bmi_c` variable.

```{r}
GGally::ggpairs(sm_oh_A_young %>% 
                    select(bmi_c, smoke100, physhealth))
```

So `bmi_c` and `smoke100` each have modest positive correlations with `physhealth` and only a very small correlation with each other. Here are some summary statistics for this final data. 

### Summary of the final subset of data

Remember that since the mean of `bmi` is `r round(mean(sm_oh_A_young$bmi),1)`, the `bmi_c` values are just `bmi` - `r round(mean(sm_oh_A_young$bmi),1)` for each subject.

```{r}
sm_oh_A_young %>% 
    select(bmi, bmi_c, smoke100, physhealth) %>%
    skim()
```

## Modeling Strategies Explored Here

We are going to predict `physhealth` using `bmi_c` and `smoke100`. 

- Remember that `physhealth` is a count of the number of poor physical health days in the past 30. 
- As a result, `physhealth` is restricted to taking values between 0 and 30. 

We will demonstrate the use of each of the following regression models, some of which are better choices than others.

1. Ordinary Least Squares (OLS) predicting `physhealth`
2. OLS predicting the logarithm of (`physhealth` + 1)
3. Poisson regression, which is appropriate for predicting counts
4. Poisson regression, adjusted to account for overdispersion
5. Negative binomial regression, also used for counts and which adjusts for overdispersion
6. Zero-inflated models, in both the Poisson and Negative Binomial varieties, which allow us to fit counts that have lots of zero values
7. A "hurdle" model, which allows us to separately fit a model to predict the incidence of "0" and then a separate model to predict the value of `physhealth` when we know it is not zero
8. Tobit regression, where a lower (and upper) bound may be set, but the underlying model describes a latent variable which can extend beyond these boundaries

### What Will We Demonstrate?

With each approach, we will fit the model and specify procedures for doing so in R. Then we will:

1. Specify the fitted model equation
2. Interpret the model's coefficient estimates and 95% confidence intervals around those estimates.
3. Perform a test of whether each variable adds value to the model, when the other one is already included.
4. Store the fitted values and appropriate residuals for each model.
5. Summarize the model's apparent R^2^ value, the proportion of variation explained, and the model log likelihood.
6. Perform checks of model assumptions as appropriate.
7. Describe how predictions would be made for two new subjects.
    - Harry has a BMI that is 10 kg/m^2^ higher than the average across all respondents and has smoked more than 100 cigarettes in his life.
    - Sally has a BMI that is 5 kg/m^2^ less than the average across all respondents and has not smoked more than 100 cigarettes in her life.

In addition, for some of the new models, we provide a little of the mathematical background, and point to other resources you can use to learn more about the model.
    
### Extra Data File for Harry and Sally    
    
To make our lives a little easier, I'll create a little tibble containing the necessary data for Harry and Sally.

```{r}
hs_data <- data_frame(subj = c("Harry", "Sally"),
                      bmi_c = c(10, -5),
                      smoke100 = c(1, 0))
hs_data
```


## The OLS Approach

```{r}
mod_ols1 <- lm(physhealth ~ bmi_c + smoke100, 
               data = sm_oh_A_young)

summary(mod_ols1)
confint(mod_ols1)
```

### The Fitted Equation

The OLS fitted model equation is 

```
physhealth = 2.00 + 0.15 bmi_c + 2.03 smoke100
```

### Interpreting the Coefficients

- The intercept, 2.00, is the predicted `physhealth` (in days) for a subject with average BMI who has not smoked 100 cigarettes or more.
- The `bmi_c` coefficient, 0.15, indicates that for each additional kg/m^2^ of BMI, while holding `smoke100` constant, the predicted `physhealth` value increases by 0.15 day.
- The `smoke100` coefficient, 2.03, indicates that a subject who has smoked 100 cigarettes or more has a predicted `physhealth` value 2.03 days larger than another subject with the same `bmi` but who has not smoked 100 cigarettes.

### Testing the Predictors

We can use the t test results provided above to assess the significance of each term, after the other is already included in the model. Each *p* value is well below our usual $\alpha$ levels, so we conclude that each predictor adds statistically detectable predictive value to the model.

### Store fitted values and residuals

We can use `broom` to do this. Here, for instance, is a table of the first six predictions and residuals.

```{r}
sm_ols_1 <- augment(mod_ols1, sm_oh_A_young)

sm_ols_1 %>% select(physhealth, .fitted, .resid) %>% head()
```

It turns out that `r sum(sm_ols_1$.fitted < 0)` of the `r nrow(sm_ols_1)` predictions that we make are below 0, and the largest prediction made by this model is `r round(max(sm_ols_1$.fitted),2)` days.

### Specify the R^2^ and log(likelihood) values

The `glance` function in the `broom` package gives us the raw and adjusted R^2^ values, and the model log(likelihood), among other summaries.

```{r}
glance(mod_ols1) %>% round(3)
```

Here, we have

Model | R^2^ | log(likelihood)
----------: | ------: | --------:
OLS | .047 | -5409.3

### Check model assumptions

Here is a plot of the residuals vs. the fitted values for this OLS model.

```{r}
ggplot(sm_ols_1, aes(x = .fitted, y = .resid)) +
    geom_point() +
    labs(title = "Residuals vs. Fitted Values for OLS model")
```

As usual, we can check OLS assumptions (linearity, homoscedasticity and normality) with R's complete set of residual plots.

```{r, fig.height = 6}
par(mfrow = c(2,2))
plot(mod_ols1)
par(mfrow = c(1,1))
```

We see the problem with our residuals. They don't follow a Normal distribution.

### Predictions for Harry and Sally

```{r}
predict(mod_ols1, newdata = hs_data,
        interval = "prediction")
```

The prediction for Harry is 5.6 days, and for Sally is 1.2 days. The prediction intervals for each include some values below 0, which is the smallest possible value.

### Notes

- This model could have been estimated using the `ols` function in the `rms` package, as well.

```{r}
dd <- datadist(sm_oh_A_young)
options(datadist = "dd")

(mod_ols1a <- ols(physhealth ~ bmi_c + smoke100,
                 data = sm_oh_A_young, x = TRUE, y = TRUE))
```


## OLS model on log(`physhealth` + 1) days

We could try to solve the problem of fitting some predictions below 0 by log-transforming the data, so as to force values to be at least 0. Since we have undefined values when we take the log of 0, we'll add one to each of the `physhealth` values before taking logs, and then transform back when we want to make predictions.

```{r}
mod_ols_log1 <- lm(log(physhealth + 1) ~ bmi_c + smoke100,
                   data = sm_oh_A_young)

summary(mod_ols_log1)
confint(mod_ols_log1)
```

### The Fitted Equation

The equation here is

```
log(physhealth + 1) = 0.48 + 0.026 bmi_c + 0.28 smoke100
```

### Interpreting the Coefficients

- The intercept, 0.48, is the predicted logarithm of (`physhealth` + 1) (in days) for a subject with average BMI who has not smoked 100 cigarettes or more.
    - We can exponentiate to see that the prediction for (`physhealth` + 1) here is `exp(0.48)` = `r round(exp(0.48),2)` so the predicted `physhealth` for a subject with average BMI who has not smoked 100 cigarettes is `r round(exp(0.48),2) - 1` days.
- The `bmi_c` coefficient, 0.15, indicates that for each additional kg/m^2^ of BMI, while holding `smoke100` constant, the predicted logarithm of (`physhealth` + 1) increases by 0.026
- The `smoke100` coefficient, 0.28, indicates that a subject who has smoked 100 cigarettes or more has a predicted log of (`physhealth` + 1) value that is 0.28 larger than another subject with the same `bmi` but who has not smoked 100 cigarettes.

### Testing the Predictors

We are still fitting an OLS model here, so we can still use the t test results provided above to assess the significance of each term, after the other is already included in the model. Each *p* value is well below our usual $\alpha$ levels, so we conclude that each predictor adds statistically detectable predictive value to the model.

### Store fitted values and residuals

We can use `broom` to help us with this. Here, for instance, is a table of the first six predictions and residuals, on the scale of our transformed response, log(`physhealth` + 1).

```{r}
sm_ols_log1 <- augment(mod_ols_log1, sm_oh_A_young)

sm_ols_log1 <- sm_ols_log1 %>% 
    mutate(outcome = log(physhealth + 1))

sm_ols_log1 %>% 
    select(physhealth, outcome, .fitted, .resid) %>%
    head()
```

Note that the `outcome` used in this model is log(`physhealth` + 1), so the `.fitted` and `.resid` values react to that outcome, and not to our original `physhealth`.

Another option would be to calculate the model-predicted `physhealth`, which I'll call *ph* for a moment, with the formula:

$$
ph = e^{.fitted} - 1
$$

```{r}
sm_ols_log1 <- sm_ols_log1 %>% 
    mutate(pred.physhealth = exp(.fitted) - 1,
           res.physhealth = physhealth - pred.physhealth)

sm_ols_log1 %>%
    select(physhealth, pred.physhealth, res.physhealth) %>% 
    head()
```

It turns out that `r sum(sm_ols_log1$pred.physhealth < 0)` of the `r nrow(sm_ols_log1)` predictions that we make are below 0, and the largest prediction made by this model is `r round(max(sm_ols_log1$pred.physhealth),2)` days.

### Specify the R^2^ and log(likelihood) values

The `glance` function in the `broom` package gives us the raw and adjusted R^2^ values, and the model log(likelihood), among other summaries.

```{r}
glance(mod_ols_log1) %>% round(3)
```

Here, we have

Model | Scale | R^2^ | log(likelihood)
----------: | ------: | --------: | ---------:
OLS on log | log(`physhealth` + 1) | .054 | -2262.0

### Getting R^2^ on the scale of `physhealth`

We could find the correlation of our model-predicted `physhealth` values, after back-transformation, and our observed `physhealth` values, if we wanted to, and then square that to get a sort of R^2^ value. But this model is not linear in `physhealth`, of course, so it's not completely comparable to our prior OLS model.

### Check model assumptions

As usual, we can check OLS assumptions (linearity, homoscedasticity and normality) with R's complete set of residual plots. Of course, these residuals and fitted values are now on the log(`physhealth` + 1) scale.

```{r, fig.height = 6}
par(mfrow = c(2,2))
plot(mod_ols_log1)
par(mfrow = c(1,1))
```

### Predictions for Harry and Sally

```{r}
predict(mod_ols_log1, newdata = hs_data, 
        interval = "prediction", type = "response")
```

Again, these predictions are on the log(`physhealth` + 1) scale, and so we have to exponentiate them, and then subtract 1, to see them on the original `physhealth` scale.

```{r}
exp(predict(mod_ols_log1, newdata = hs_data, 
            interval = "prediction", type = "response")) - 1
```

The prediction for Harry is now 1.77 days, and for Sally is 0.41 days. The prediction intervals for each again include some values below 0, which is the smallest possible value.

## A Poisson Regression Model

The `physhealth` data describe a count. Specifically a count of the number of days where the subject felt poorly in the last 30. Why wouldn't we model this count with linear regression?

- A count can only be positive. Linear regression would estimate some subjects as having negative counts.
- A count is unlikely to follow a Normal distribution. In fact, it's far more likely that the log of the counts will follow a Poisson distribution.

So, we'll try that. The Poisson distribution is used to model a *count* outcome - that is, an outcome with possible values (0, 1, 2, ...). The model takes a somewhat familiar form to the models we've used for linear and logistic regression^[This discussion is motivated by Section 6.2 of Gelman and Hill.]. If our outcome is *y* and our linear predictors *X*, then the model is:

$$
y_i \sim \mbox{Poisson}(\lambda_i)
$$

The parameter $\lambda$ must be positive, so it makes sense to fit a linear regression on the logarithm of this...

$$
\lambda_i = exp(\beta_0 + \beta_1 X_1 + ... \beta_k X_k)
$$

The coefficients $\beta$ can be exponentiated and treated as multiplicative effects. 

We'll run a generalized linear model with a log link function, ensuring that all of the predicted values will be positive, and using a Poisson error distribution. This is called **Poisson regression**.

Poisson regression may be appropriate when the dependent variable is a count of events. The events must be independent - the occurrence of one event must not make any other more or less likely. That's hard to justify in our case, but we can take a look.

```{r}
mod_poiss1 <- glm(physhealth ~ bmi_c + smoke100, 
                  family = poisson(),
                  data = sm_oh_A_young)

summary(mod_poiss1)
confint(mod_poiss1)
```

### The Fitted Equation

The model equation is 

```
log(physhealth) = 0.63 + 0.043 bmi_c + 0.71 smoke100
```

It looks like both `bmi` and `smoke_100` have confidence intervals excluding 0.

### Interpreting the Coefficients

Our new model for $y_i$ = counts of poor `physhealth` days in the last 30, follows the regression equation:

$$
y_i \sim \mbox{Poisson}(exp(0.63 + 0.043 bmi_c + 0.71 smoke100))
$$

where `smoke100` is 1 if the subject has smoked 100 cigarettes (lifetime) and 0 otherwise, and `bmi_c` is just the centered body-mass index value in kg/m^2^. We interpret the coefficients as follows:

- The constant term, 0.63, gives us the intercept of the regression - the prediction if `smoke100 = 0` and `bmi_c = 0`. In this case, because we've centered BMI, it implies that `exp(0.63)` = `r round(exp(0.63),2)` is the predicted days of poor `physhealth` for a non-smoker with average BMI. 
- The coefficient of `bmi_c`, 0.043, is the expected difference in count of poor `physhealth` days (on the log scale) for each additional kg/m^2^ of body mass index. The expected multiplicative *increase* is $e^{0.043}$ = `r round(exp(0.043),3)`, corresponding to a 4.4% difference in the count.
- The coefficient of `smoke100`, 0.71, tells us that the predictive difference between those who have and who have not smoked 100 cigarettes can be found by multiplying the `physhealth` count by exp(0.71) = `r round(exp(0.71),2)`, yielding essentially a doubling of the `physhealth` count.

As with linear or logistic regression, each coefficient is interpreted as a comparison where one predictor changes by one unit, while the others remain constant.

### Testing the Predictors

We can use the Wald tests (z tests) provided with the Poisson regression output, or we can fit the model and then run an ANOVA to obtain a test based on the deviance (a simple transformation of the log likelihood ratio.) 

- By the Wald tests shown above, each predictor clearly adds significant predictive value to the model given the other predictor, and we note that the *p* values are as small as R will support.
- The ANOVA approach for this model lets us check the impact of adding `smoke100` to a model already containing `bmi_c`.

```{r}
anova(mod_poiss1, test = "Chisq")
```

To obtain a *p* value for `smoke100`'s impact after `bmi_c` is accounted for, we compare the difference in deviance to a chi-square distribution with 1 degree of freedom. To check the effect of `bmi_c`, we could refit the model with `bmi_c` entering last, and again run an ANOVA.

We could also run a likelihood-ratio test for each predictor, by fitting the model with and without that predictor.

```{r}
mod_poiss1_without_bmi <- glm(physhealth ~ smoke100,
                              family = poisson(),
                              data = sm_oh_A_young)

anova(mod_poiss1, mod_poiss1_without_bmi, test = "Chisq")
```

### Correcting for Overdispersion with `coeftest`/`coefci`

The main assumption we'll think about in a Poisson model is about **overdispersion**. We might deal with the overdispersion we see in this model by changing the nature of the tests we run within this model, using the `coeftest` or `coefci` approaches from the `lmtest` package, as I'll demonstrate next, or we might refit the model using a quasi-likelihood approach, as I'll show in the material to come.

Here, we'll use the `coeftest` and `coefci` approach from `lmtest` combined with robust sandwich estimation (via the `sandwich` package) to re-compute the Wald tests.

```{r}
coeftest(mod_poiss1, vcov. = sandwich)
coefci(mod_poiss1, vcov. = sandwich)
```

Both predictors are still significant, but the standard errors are more appropriate. Later, we'll fit this approach by changing the estimation method to a quasi-likelihood approach.

### Store fitted values and residuals

What happens if we try using the `broom` package in this case? We can, if we like, get our residuals and predicted values right on the scale of our `physhealth` response.

```{r}
sm_poiss1 <- augment(mod_poiss1, sm_oh_A_young,
                     type.predict = "response",
                     type.residuals = "response")

sm_poiss1 %>% 
    select(physhealth, .fitted, .resid) %>%
    head()
```

### Rootogram: see the fit of a count regression model

A **rootogram** is a very useful way to visualize the fit of a count regression model^[See http://data.library.virginia.edu/getting-started-with-negative-binomial-regression-modeling/]. The `rootogram` function in the `countreg` package makes this pretty straightforward. By default, this fits a hanging rootogram on the square root of the frequencies.

```{r}
rootogram(mod_poiss1, max = 30)
```

The red curved line is the theoretical Poisson fit. "Hanging" from each point on the red line is a bar, the height of which represents the difference between expected and observed counts. A bar hanging below 0 indicates underfitting. A bar hanging above 0 indicates overfitting. The counts have been transformed with a square root transformation to prevent smaller counts from getting obscured and overwhelmed by larger counts. We see a great deal of underfitting for counts of 0, and overfitting for most other counts, especially 1-6, with some underfitting again by `physhealth` above 14 days.

### Specify the R^2^ and log(likelihood) values

We can calculate the R^2^ as the squared correlation of the fitted values and the observed values.

```{r}
# The correlation of observed and fitted values
(poiss_r <- with(sm_poiss1, cor(physhealth, .fitted)))

# R-square
poiss_r^2
```

The `glance` function in the `broom` package gives us model log(likelihood), among other summaries.

```{r}
glance(mod_poiss1) %>% round(3)
```

Here, we have

Model | Scale | R^2^ | log(likelihood)
----------: | ------: | --------: | ---------:
Poisson | log(`physhealth`) | .049 | -7841.69

### Check model assumptions

The Poisson model is a classical generalized linear model, estimated using the method of maximum likelihood. While the default `plot` option for a `glm` still shows the plots we would use to assess the assumptions of an OLS model, we don't actually get much from that, since our Poisson model has different assumptions. It can be useful to look at a plot of residuals vs. fitted values on the original `physhealth` scale.

```{r}
ggplot(sm_poiss1, aes(x = .fitted, y = .resid)) +
    geom_point() +
    labs(title = "Residuals vs. Fitted `physhealth`",
         subtitle = "Original Poisson Regression model")
```

### Using `glm.diag.plots` from the `boot` package

The `glm.diag.plots` function from the `boot` package makes a series of diagnostic plots for generalized linear models.

- (Top, Left) Jackknife deviance residuals against fitted values. This is essentially identical to what you obtain with `plot(mod_poiss1, which = 1)`. A *jackknife deviance* residual is also called a likelihood residual. It is the change in deviance when this observation is omitted from the data. 
- (Top, Right) Normal Q-Q plot of standardized deviance residuals. (Dotted line shows expectation if those standardized residuals followed a Normal distribution, and these residuals generally should.) The result is similar to what you obtain with `plot(mod_poiss1, which = 2)`.
- (Bottom, Left) Cook statistic vs. standardized leverage 
    + n = # of observations, p = # of parameters estimated
    + Horizontal dotted line is at $\frac{8}{n - 2p}$. Points above the line have high influence on the model.
    + Vertical line is at $\frac{2p}{n - 2p}$. Points to the right of the line have high leverage.
- (Bottom, Right) Index plot of Cook's statistic to help identify the observations with high influence. This is essentially the same plot as `plot(mod_poiss1, which = 4)`

```{r, fig.height = 6}
glm.diag.plots(mod_poiss1)
```

When working with these plots, it is possible to use the `iden` command to perform some interactive identification of points in your R terminal. But that doesn't play out effectively in an HTML summary document like this, so we'll leave that out.

### Predictions for Harry and Sally

The predictions from a `glm` fit like this don't include prediction intervals. But we can get predictions on the scale of our original response variable, `physhealth`, like this.

```{r}
predict(mod_poiss1, newdata = hs_data, se.fit = TRUE,
        type = "response")
```

By using `response` as the type, these predictions fall on the original `physhealth` scale. The prediction for Harry is now 5.87 days, and for Sally is 1.52 days. 

## Overdispersion in a Poisson Model

Poisson regressions do not supply an independent variance parameter $\sigma$, and as a result can be overdispersed, and usually are. Under the Poisson distribution, the variance equals the mean - so the standard deviation equals the square root of the mean. The notion of **overdispersion** arises here. When fitting generalized linear models with Poisson error distributions, the residual deviance and its degrees of freedom should be approximately equal if the model fits well.

If the residual deviance is far greater than the degrees of freedom, then overdispersion may well be a problem. In this case, the residual deviance is about 8.5 times the size of the residual degrees of freedom, so that's a clear indication of overdispersion. We saw earlier that the Poisson regression model requires that the outcome (here the `physhealth` counts) be independent. A possible reason for the overdispersion we see here is that `physhealth` on different days likely do not occur independently of one another but likely "cluster" together.

### Testing for Overdispersion?

Gelman and Hill provide an overdispersion test in R for a Poisson model as follows...

```{r}
yhat <- predict(mod_poiss1, type = "response")
n <- arm::display(mod_poiss1)$n
k <- arm::display(mod_poiss1)$k
z <- (sm_oh_A_young$physhealth - yhat) / sqrt(yhat)
cat("overdispersion ratio is ", sum(z^2)/ (n - k), "\n")
cat("p value of overdispersion test is ", 
    pchisq(sum(z^2), df = n-k, lower.tail = FALSE), "\n")
```

The p value here is 0, indicating that the probability is essentially zero that a random variable from a $\chi^2$ distribution with (n - k) = 1633 degrees of freedom would be as large as what we observed in this case. So there is significant overdispersion.

In summary, the `physhealth` counts are overdispersed by a factor of 15.499, which is enormous (even a factor of 2 would be considered large) and also highly statistically significant. The basic correction for overdisperson is to multiply all regression standard errors by $\sqrt{15.499}$ = `r round(sqrt(15.499),2)`. 

The `quasipoisson` model and the negative binomial model that we'll fit below are very similar. We write the overdispersed "quasiPoisson" model as:

$$
y_i \sim \mbox{overdispersed Poisson} (\mu_i exp(X_i \beta), \omega)
$$

where $\omega$ is the overdispersion parameter, 15.499, in our case. The Poisson model we saw previously is then just the overdispersed Poisson model with $\omega = 1$. 

## Fitting the Quasi-Poisson Model

To deal with overdispersion, one useful approach is to apply a **quasi-likelihood estimation procedure**, as follows:

```{r}
mod_poiss_od1 <- glm(physhealth ~ bmi_c + smoke100, 
                  family = quasipoisson(),
                  data = sm_oh_A_young)

summary(mod_poiss_od1)
confint(mod_poiss_od1)
```

This "quasi-Poisson regression" model uses the same mean function as Poisson regression, but now estimated by quasi-maximum likelihood estimation or, equivalently, through the method of generalized estimating equations, where the inference is adjusted by an estimated dispersion parameter. Sometimes, though I won't demonstrate this here, people fit an "adjusted" Poisson regression model, where this estimation by quasi-ML is augmented to adjust the inference via sandwich estimates of the covariances^[See Zeileis A Kleiber C Jackman S *Regression Models for Count Data in R* Vignette at https://cran.r-project.org/web/packages/pscl/vignettes/countreg.pdf].

### The Fitted Equation

The model equation is still `log(physhealth) = 0.63 + 0.04 bmi_c + 0.71 smoke100`. The estimated coefficients are still statistically significant, but the standard errors for each coefficient are considerably larger when we account for overdispersion.

The dispersion parameter for the quasi-Poisson family is now taken to be a bit less than the square root of the ratio of the residual deviance and its degrees of freedom. This is a much more believable model, as a result. 

### Interpreting the Coefficients

No meaningful change from the Poisson model we saw previously.

- The constant term, 0.63, gives us the intercept of the regression - the prediction if `smoke100 = 0` and `bmi_c = 0`. In this case, because we've centered BMI, it implies that `exp(0.63)` = `r round(exp(0.63),2)` is the predicted days of poor `physhealth` for a non-smoker with average BMI. 
- The coefficient of `bmi_c`, 0.043, is the expected difference in count of poor `physhealth` days (on the log scale) for each additional kg/m^2^ of body mass index. The expected multiplicative *increase* is $e^{0.043}$ = `r round(exp(0.043),3)`, corresponding to a 4.4% difference in the count.
- The coefficient of `smoke100`, 0.71, tells us that the predictive difference between those who have and who have not smoked 100 cigarettes can be found by multiplying the `physhealth` count by exp(0.71) = `r round(exp(0.71),2)`, yielding essentially a doubling of the `physhealth` count.

### Testing the Predictors

Again, we can use the Wald tests (z tests) provided with the Poisson regression output, or we can fit the model and then run an ANOVA to obtain a test based on the deviance (a simple transformation of the log likelihood ratio.) 

- By the Wald tests shown above, each predictor clearly adds significant predictive value to the model given the other predictor, and we note that the *p* values are as small as R will support.
- The ANOVA approach for this model lets us check the impact of adding `smoke100` to a model already containing `bmi_c`.

```{r}
anova(mod_poiss_od1, test = "Chisq")
```

The result is unchanged. To obtain a *p* value for `smoke100`'s impact after `bmi_c` is accounted for, we compare the difference in deviance to a chi-square distribution with 1 degree of freedom. The result is incredibly statistically significant.

To check the effect of `bmi_c`, we could refit the model with and without `bmi_c`, and again run an ANOVA. I'll skip that here.

### Store fitted values and residuals

What happens if we try using the `broom` package in this case? We can, if we like, get our residuals and predicted values right on the scale of our `physhealth` response.

```{r}
sm_poiss_od1 <- augment(mod_poiss_od1, sm_oh_A_young,
                     type.predict = "response",
                     type.residuals = "response")

sm_poiss_od1 %>% 
    select(physhealth, .fitted, .resid) %>%
    head()
```

It turns out that `r sum(sm_poiss_od1$.fitted < 0)` of the `r nrow(sm_poiss_od1)` predictions that we make are below 0, and the largest prediction made by this model is `r round(max(sm_poiss_od1$.fitted),2)` days.

The `rootogram` function we've shown doesn't support overdispersed Poisson models at the moment.

### Specify the R^2^ and log(likelihood) values

We can calculate the R^2^ as the squared correlation of the fitted values and the observed values.

```{r}
# The correlation of observed and fitted values
(poiss_od_r <- with(sm_poiss_od1, cor(physhealth, .fitted)))

# R-square
poiss_od_r^2
```

The `glance` function in the `broom` package gives us model log(likelihood), among other summaries.

```{r}
glance(mod_poiss_od1) %>% round(3)
```

Here, we have

Model | Scale | R^2^ | log(likelihood)
----------: | ------: | --------: | ---------:
Poisson | log(`physhealth`) | .049 | NA

### Check model assumptions

Having dealt with the overdispersion, this should be a cleaner model in some ways, but the diagnostics (other than the dispersion) will be the same. Here is a plot of residuals vs. fitted values on the original `physhealth` scale.

```{r}
ggplot(sm_poiss_od1, aes(x = .fitted, y = .resid)) +
    geom_point() +
    labs(title = "Residuals vs. Fitted `physhealth`",
         subtitle = "Overdispersed Poisson Regression model")
```

I'll skip the `glm.diag.plots` results, since you've already seen them. 

### Predictions for Harry and Sally

The predictions from this overdispersed Poisson regression will match those in the original Poisson regression, but the standard error will be larger.

```{r}
predict(mod_poiss_od1, newdata = hs_data, se.fit = TRUE,
        type = "response")
```

By using `response` as the type, these predictions fall on the original `physhealth` scale. Again, the prediction for Harry is 5.87 days, and for Sally is 1.52 days. 

## Poisson and Quasi-Poisson models using `Glm` from the `rms` package

The `Glm` function in the `rms` package can be used to fit both the original Poisson regression and the quasi-Poisson model accounting for overdispersion.

### Refitting the original Poisson regression with `Glm`

```{r}
d <- datadist(sm_oh_A_young)
options(datadist = "d")

mod_poi_Glm_1 <- Glm(physhealth ~ bmi_c + smoke100,
                     family = poisson(), 
                     data = sm_oh_A_young, 
                     x = T, y = T)

mod_poi_Glm_1
```

### Refitting the overdispersed Poisson regression with `Glm`

```{r}
d <- datadist(sm_oh_A_young)
options(datadist = "d")

mod_poi_od_Glm_1 <- Glm(physhealth ~ bmi_c + smoke100,
                     family = quasipoisson(), 
                     data = sm_oh_A_young, 
                     x = T, y = T)

mod_poi_od_Glm_1
```

The big advantage here is that we have access to the usual `ANOVA`, `summary`, and `nomogram` features that `rms` brings to fitting models.

### ANOVA on a `Glm` fit

```{r}
anova(mod_poi_od_Glm_1)
```

This shows the individual Wald $\chi^2$ tests without having to refit the model.

### ggplots from `Glm` fit

```{r}
ggplot(Predict(mod_poi_od_Glm_1, fun = exp))
```

### Summary of a `Glm` fit

```{r}
summary(mod_poi_od_Glm_1)
```

### Plot of the Summary

```{r}
plot(summary(mod_poi_od_Glm_1))
```

### Nomogram of a `Glm` fit

```{r}
plot(nomogram(mod_poi_od_Glm_1, fun = exp, 
              funlabel = "physhealth days"))
```

Note the use of `fun=exp` in both the ggplot of Predict and the nomogram. What's that doing?


## Negative Binomial Model

Another approach to dealing with overdispersion is to fit a negative binomial model^[See https://cran.r-project.org/web/packages/pscl/vignettes/countreg.pdf for more details.] to predict the log(`physhealth`) counts. This involves the fitting of an additional parameter, $\theta$. That's our dispersion parameter^[This $\theta$ is the inverse of the dispersion parameter estimated for these models by most other software packages, like SAS, Stata and SPSS. See https://stats.idre.ucla.edu/r/dae/negative-binomial-regression/ for more details.] 

Sometimes, people will fit a model where $\theta$ is known, for instance a geometric model (where $\theta$ = 1), and then this can be directly plugged into a `glm()` fit, but the more common scenario is that we are going to iteratively estimate the $\beta$ coefficients and $\theta$. To do this, I'll use the `glm.nb` function from the `MASS` package.

```{r}
mod_nb1 <- MASS::glm.nb(physhealth ~ bmi_c + smoke100, link = log,
                  data = sm_oh_A_young)

summary(mod_nb1)
confint(mod_nb1)
```

### The Fitted Equation

The form of the model equation for a negative binomial regression is the same as that for Poisson regression. 

```
log(physhealth) = 0.61 + 0.043 bmi_c + 0.756 smoke100
```

### Comparison with the (raw) Poisson model

To compare the negative binomial model to the Poisson model (without the overdispersion) we can use the `logLik` function to make a comparison. Note that the Poisson model is a subset of the negative binomial.

```{r}
logLik(mod_nb1)
logLik(mod_poiss1)
2 * (logLik(mod_nb1) - logLik(mod_poiss1))
     
pchisq(2 * (logLik(mod_nb1) - logLik(mod_poiss1)), df = 1, lower.tail = FALSE)
```

Here, the difference in the log likelihoods is large enough that the resulting *p* value is very small. This strongly suggests that the negative binomial model, which adds the dispersion parameter, is more appropriate than the raw Poisson model.

However, both the regression coefficients and the standard errors are rather similar to the quasi-Poisson and the sandwich-adjusted Poisson results above. Thus, in terms of predicted means, all three models give very similar results; the associated Wald tests also lead to the same conclusions.

### Interpreting the Coefficients

There's only a small change here from the Poisson models we saw previously.

- The constant term, 0.61, gives us the intercept of the regression - the prediction if `smoke100 = 0` and `bmi_c = 0`. In this case, because we've centered BMI, it implies that `exp(0.61)` = `r round(exp(0.61),2)` is the predicted days of poor `physhealth` for a non-smoker with average BMI. 
- The coefficient of `bmi_c`, 0.043, is the expected difference in count of poor `physhealth` days (on the log scale) for each additional kg/m^2^ of body mass index. The expected multiplicative *increase* is $e^{0.043}$ = `r round(exp(0.043),3)`, corresponding to a 4.4% difference in the count.
- The coefficient of `smoke100`, 0.756, tells us that the predictive difference between those who have and who have not smoked 100 cigarettes can be found by multiplying the `physhealth` count by exp(0.756) = `r round(exp(0.756),2)`, yielding essentially a doubling of the `physhealth` count.

### Interpretation of Coefficients in terms of IRRs

We might be interested in looking at incident rate ratios rather than coefficients. The coefficients have an additive effect in the log(y) scale, and the IRR have a multiplicative effect in the y scale. To do this, we can exponentiate our model coefficients. This also applies to the confidence intervals.

```{r}
exp(coef(mod_nb1))
exp(confint(mod_nb1))
```

As an example, then, the incident rate for `smoke100` = 1 is 2.13 times the incident rate of `physhealth` days for the reference group (`smoke100` = 0). The percent change in the incident rate of `physhealth` is a 4.3% increase for every kg/m^2^ increase in centered `bmi`.

### Testing the Predictors

Again, we can use the Wald tests (z tests) provided with the negative binomial regression output. 

As an alternative, we probably should not use the standard `anova` process, because the models there don't re-estimate $\theta$ for each new model, as the warning message below indicates.

```{r}
anova(mod_nb1)
```

So, instead, if we want, for instance to assess the significance of `bmi_c`, after `smoke100` is already included in the model, we fit both models (with and without `bmi_c`) and then compare those models with a likelihood ratio test.

```{r}
mod_nb1_without_bmi <- MASS::glm.nb(physhealth ~ smoke100, 
                                      link = log,
                                      data = sm_oh_A_young)

anova(mod_nb1, mod_nb1_without_bmi)
```

And we could compare the negative binomial models with and without `smoke100` in a similar way.

```{r}
mod_nb1_without_smoke <- MASS::glm.nb(physhealth ~ bmi_c, 
                                      link = log,
                                      data = sm_oh_A_young)

anova(mod_nb1, mod_nb1_without_smoke)
```

### Store fitted values and residuals

The `broom` package works in this case, too. We'll look here at residuals and predicted (fitted) values on the scale of our `physhealth` response.

```{r}
sm_nb1 <- augment(mod_nb1, sm_oh_A_young,
                     type.predict = "response",
                     type.residuals = "response")

sm_nb1 %>% 
    select(physhealth, .fitted, .resid) %>%
    head()
```

### Rootogram for Negative Binomial model

Here's the rootogram for the negative binomial model.

```{r}
rootogram(mod_nb1, max = 30)
```

Again, the red curved line is the theoretical (negative binomial) fit. "Hanging" from each point on the red line is a bar, the height of which represents the difference between expected and observed counts. A bar hanging below 0 indicates underfitting. A bar hanging above 0 indicates overfitting. The counts have been transformed with a square root transformation to prevent smaller counts from getting obscured and overwhelmed by larger counts. 

The match looks much better than the Poisson model, which is a sign that accounting for overdispersion is very important. Even this model badly underfits the number of 30 values, however.

### Simulating what the Negative Binomial model predicts

We can use the parameters of the negative binomial model to simulate data^[See http://data.library.virginia.edu/getting-started-with-negative-binomial-regression-modeling/] and compare the simulated results to our observed `physhealth` data.

```{r}
par(mfrow=c(1,2))
sm_oh_A_young$physhealth %>% 
    table() %>% barplot(main = "Observed physhealth")
set.seed(432122)
rnbinom(n = nrow(sm_oh_A_young), 
        size = mod_nb1$theta,
        mu = exp(coef(mod_nb1)[1])) %>%
    table() %>% barplot(main = "Simulated physhealth")
```

Again we see that the simulated data badly underfits the 30 values, and includes some predictions larger than 30.

### Specify the R^2^ and log(likelihood) values

We can calculate the R^2^ as the squared correlation of the fitted values and the observed values.

```{r}
# The correlation of observed and fitted values
(nb_r <- with(sm_nb1, cor(physhealth, .fitted)))

# R-square
nb_r^2
```

The `glance` function in the `broom` package gives us model log(likelihood), among other summaries.

```{r}
glance(mod_nb1) %>% round(3)
```

Here, we have

Model | Scale | R^2^ | log(likelihood)
----------: | ------: | --------: | ---------:
Negative Binomial | log(`physhealth`) | .049 | -2585.6

### Check model assumptions

Here is a plot of residuals vs. fitted values on the original `physhealth` scale.

```{r}
ggplot(sm_nb1, aes(x = .fitted, y = .resid)) +
    geom_point() +
    labs(title = "Residuals vs. Fitted `physhealth`",
         subtitle = "Negative Binomial Regression model")
```

Here are the `glm` diagnostic plots from the `boot` package.

```{r, fig.height = 6}
glm.diag.plots(mod_nb1)
```

From the lower left plot, we see fewer points with large values of both Cook's distance and leverage, so that's a step in the right direction. The upper right plot still has some issues, but we're closer to a desirable result there, too.

### Predictions for Harry and Sally

The predictions from this negative binomial regression model will be only a little different than those from the Poisson models.

```{r}
predict(mod_nb1, newdata = hs_data, se.fit = TRUE,
        type = "response")
```

As we've seen in the past, when we use `response` as the type, the predictions fall on the original `physhealth` scale. The prediction for Harry is 5.99 days, and for Sally is 1.49 days. 

## The Problem: Too Few Zeros

Remember that we observe more than 1000 zeros in our `physhealth` data.

```{r}
sm_oh_A_young %>% count(physhealth == 0)
```

Let's go back to our Poisson model (without overdispersion) for a moment, and concentrate on the zero values. 

```{r}
# predict expected mean physhealth for each subject
mu <- predict(mod_poiss1, type = "response")

# sum the probabilities of a zero count for each mean
exp <- sum(dpois(x = 0, lambda = mu))

# predicted number of zeros from Poisson model
round(exp)
```

As we've seen previously, we're severely underfitting zero counts. We can compare the observed number of zero `physhealth` results to the expected number of zero values from the likelihood-based models.

```{r}
round(c("Obs" = sum(sm_oh_A_young$physhealth == 0),
  "Poisson" = sum(dpois(0, fitted(mod_poiss1))),
  "NB" = sum(dnbinom(0, mu = fitted(mod_nb1), size = mod_nb1$theta))),0)
```

There are at least two ways to tackle this problem.

- Fitting a model which deliberately inflates the number of zeros that are fitted
- Fitting a hurdle model

We'll look at those options, next.

## The Zero-Inflated Poisson Regression Model

The zero-inflated Poisson or (ZIP) model is used to describe count data with an excess of zero counts^[See https://stats.idre.ucla.edu/r/dae/zip/ for more on the zero-inflated poisson model.]. The model posits that there are two processes involved:

- a logit model is used to predict excess zeros
- while a Poisson model is used to predict the counts, generally

The `pscl` package is used here, which can conflict with the `countreg` package we used to fit rootograms. That's why I'm loading it here.

```{r}
library(pscl)
```

To run the zero-inflated Poisson model, we use the following:

```{r}
mod_zip1 <- zeroinfl(physhealth ~ bmi_c + smoke100, 
                     data = sm_oh_A_young)

summary(mod_zip1)
confint(mod_zip1)
```

The output describes two separate regression models. Below the model call, we see information on a Poisson regression model. Then we see another block describing the inflation model.

Each predictor (`bmi_c` and `smoke100`) appears to be statistically significant in each part of the model.

### Comparison to a null model

To show that this model fits better than the null model (the model with intercept only), we can compare them directly with a chi-squared test. Since we have two predictors in the full model, the degrees of freedom for this test is 2.

```{r}
mod_zipnull <- zeroinfl(physhealth ~ 1, 
                     data = sm_oh_A_young)

summary(mod_zipnull)

pchisq(2 * (logLik(mod_zip1) - logLik(mod_zipnull)), df = 2, lower.tail = FALSE)
```

### Comparison to a Poisson Model with the Vuong test

```{r}
vuong(mod_zip1, mod_poiss1)
```

Certainly, the ZIP model is a significant improvement over the standard Poisson model, as the Vuong test reveals.

### The Fitted Equation

The form of the model equation for a zero-inflated Poisson regression requires us to take two separate models into account. First we have a logistic regression model to predict the log odds of zero `physhealth` days...

```
logit(physhealth = 0) = 0.919 - 0.051 bmi_c - 0.411 smoke100
```

That takes care of the *extra* zeros. Then, to predict the number of `physhealth` days, we have:

```
log(physhealth) = 1.90 + 0.014 bmi_c + 0.441 smoke100
```

which may produce some additional zero count estimates.

### Interpreting the Coefficients

We can exponentiate the logistic regression coefficients to obtain results in terms of odds ratios for that model, and that can be of some help in understanding the process behind excess zeros.

Also, exponentiating the coefficients of the count model help us describe those counts on the original scale of `physhealth`.

```{r}
exp(coef(mod_zip1))
```

For example,

- in the model for `physhealth` = 0, the odds of `physhealth` = 0 are 66% as high for subjects with `smoke100` = 1 as for non-smokers with the same BMI.
- in the Poisson model for `physhealth`, the `physhealth` count is estimated to increase by 1.55 for smokers as compared to non-smokers with the same BMI. 

### Testing the Predictors

We can test the model with and without `bmi_c`, for example, by fitting the model both ways, and comparing the results with either a Wald or Likelihood Ratio test, each of which is available in the `lmtest` package.

```{r}
mod_zip1_nobmi <- zeroinfl(physhealth ~ smoke100, 
                     data = sm_oh_A_young)

lmtest::waldtest(mod_zip1, mod_zip1_nobmi)
```

```{r}
lmtest::lrtest(mod_zip1, mod_zip1_nobmi)
```

### Store fitted values and residuals

The `broom` package does not work with the `zeroinfl` tool. So we need to build up the fitted values and residuals ourselves.

```{r}
sm_zip1 <- sm_oh_A_young %>%
    mutate(fitted = fitted(mod_zip1, type = "response"),
           resid = resid(mod_zip1, type = "response"))

sm_zip1 %>% 
    dplyr::select(physhealth, fitted, resid) %>%
    head()
```

### Modeled Number of Zero Counts

The zero-inflated model is designed to perfectly match the number of observed zeros. We can compare the observed number of zero `physhealth` results to the expected number of zero values from the likelihood-based models.

```{r}
round(c("Obs" = sum(sm_oh_A_young$physhealth == 0),
  "Poisson" = sum(dpois(0, fitted(mod_poiss1))),
  "NB" = sum(dnbinom(0, mu = fitted(mod_nb1), size = mod_nb1$theta)),
  "ZIP" = sum(predict(mod_zip1, type = "prob")[,1])), 0)
```

### Rootogram for ZIP model

Here's the rootogram for the zero-inflated Poisson model.

```{r}
countreg::rootogram(mod_zip1, max = 30)
```

The zero frequencies are perfectly matched here, but we can see that counts of 1 and 2 are now substantially underfit, and values between 6 and 13 are overfit.

### Specify the R^2^ and log (likelihood) values

We can calculate a proxy for R^2^ as the squared correlation of the fitted values and the observed values.

```{r}
# The correlation of observed and fitted values
(zip_r <- with(sm_zip1, cor(physhealth, fitted)))

# R-square
zip_r^2
```

```{r}
logLik(mod_zip1)
```

Here, we have

Model | Scale | R^2^ | log(likelihood)
----------: | ------: | --------: | ---------:
Zero-Inflated Poisson | Complex: log(`physhealth`) | .050 | -4184.1

### Check model assumptions

Here is a plot of residuals vs. fitted values on the original `physhealth` scale.

```{r}
ggplot(sm_zip1, aes(x = fitted, y = resid)) +
    geom_point() +
    labs(title = "Residuals vs. Fitted `physhealth`",
         subtitle = "Zero-Inflated Poisson Regression model")
```

### Predictions for Harry and Sally

The predictions from this ZIP regression model are obtained as follows...

```{r}
predict(mod_zip1, newdata = hs_data, type = "response")
```

As we've seen in the past, when we use `response` as the type, the predictions fall on the original `physhealth` scale. The prediction for Harry is 5.92 days, and for Sally is 1.47 days. 


## The Zero-Inflated Negative Binomial Regression Model

As an alternative to the ZIP model, we might consider a zero-inflated negative binomial regression^[See https://stats.idre.ucla.edu/r/dae/zinb/]. This will involve a logistic regression to predict the probability of a 0, and then a negative binomial model to describe the counts of `physhealth`.

To run the zero-inflated negative binomial model, we use the following code:

```{r}
mod_zinb1 <- zeroinfl(physhealth ~ bmi_c + smoke100, 
                      dist = "negbin", data = sm_oh_A_young)

summary(mod_zinb1)
confint(mod_zinb1)
```

### Comparison to a null model

To show that this model fits better than the null model (the model with intercept only), we can compare them directly with a chi-squared test. Since we have two predictors in the full model, the degrees of freedom for this test is 2.

```{r}
mod_zinbnull <- zeroinfl(physhealth ~ 1, dist = "negbin",
                     data = sm_oh_A_young)

summary(mod_zinbnull)

pchisq(2 * (logLik(mod_nb1) - logLik(mod_zinbnull)), df = 2, lower.tail = FALSE)
```

### Comparison to a Negative Binomial Model: Vuong test

```{r}
vuong(mod_zinb1, mod_nb1)
```

The zero-inflated negative binomial model is a significant improvement over the standard negative binomial model according to the the raw or AIC-corrected Vuong tests, but not according to the BIC-corrected test.

### The Fitted Equation

Like the ZIP, the zero-inflated negative binomial regression also requires us to take two separate models into account. First we have a logistic regression model to predict the log odds of zero `physhealth` days...

```
logit(physhealth = 0) = 0.044 - 0.062 bmi_c - 0.337 smoke100
```

That takes care of the *extra* zeros. Then, to predict the number of `physhealth` days, we have:

```
log(physhealth) = 1.36 + 0.020 bmi_c + 0.575 smoke100
```
and there's a $\theta$ term, which is estimated as `exp(-1.023)` or `r round(exp(-1.023),3)`, and this negative binomial regression model may also produce some additional zero count estimates.

### Interpreting the Coefficients

As with the zip, we can exponentiate the logistic regression coefficients to obtain results in terms of odds ratios for that model, and that can be of some help in understanding the process behind excess zeros.

```{r}
exp(coef(mod_zinb1))
```

For example,

- in the model for `physhealth` = 0, the odds of `physhealth` = 0 are 71.4% as high for subjects with `smoke100` = 1 as for non-smokers with the same BMI.

Interpreting the negative binomial piece works the same way as it did in the negative binomial regression.

### Testing the Predictors

We can test the model with and without `bmi_c`, for example, by fitting the model both ways, and comparing the results with either a Wald or Likelihood Ratio test, each of which is available in the `lmtest` package.

```{r}
mod_zinb1_nobmi <- zeroinfl(physhealth ~ smoke100, 
                            dist = "negbin",
                            data = sm_oh_A_young)

lmtest::waldtest(mod_zinb1, mod_zinb1_nobmi)
```

```{r}
lmtest::lrtest(mod_zinb1, mod_zinb1_nobmi)
```

### Store fitted values and residuals

Again, we need to build up the fitted values and residuals without the `broom` package.

```{r}
sm_zinb1 <- sm_oh_A_young %>%
    mutate(fitted = fitted(mod_zinb1, type = "response"),
           resid = resid(mod_zinb1, type = "response"))

sm_zip1 %>% 
    dplyr::select(physhealth, fitted, resid) %>%
    head()
```

### Modeled Number of Zero Counts

Once again, we can compare the observed number of zero `physhealth` results to the expected number of zero values from the likelihood-based models.

```{r}
round(c("Obs" = sum(sm_oh_A_young$physhealth == 0),
  "Poisson" = sum(dpois(0, fitted(mod_poiss1))),
  "NB" = sum(dnbinom(0, mu = fitted(mod_nb1), size = mod_nb1$theta)),
  "ZIP" = sum(predict(mod_zip1, type = "prob")[,1]),
  "ZINB" = sum(predict(mod_zinb1, type = "prob")[,1])),0)
```

So, the Poisson model is clearly inappropriate, but the zero-inflated (Poisson and NB) and the negative binomial model all give reasonable fits in this regard.

### Rootogram for Zero-Inflated Negative Binomial model

Here's the rootogram for the zero-inflated negative binomial model.

```{r}
countreg::rootogram(mod_zinb1, max = 30)
```

As in the ZIP model, the zero frequencies are perfectly matched here, but we can see that counts of 1 and 2 are now closer to the data we observe than in the ZIP model. We are still substantially underfitting values of 30.

### Specify the R^2^ and log (likelihood) values

We can calculate a proxy for R^2^ as the squared correlation of the fitted values and the observed values.

```{r}
# The correlation of observed and fitted values
(zinb_r <- with(sm_zinb1, cor(physhealth, fitted)))

# R-square
zinb_r^2
```

```{r}
logLik(mod_zinb1)
```

Here, we have

Model | Scale | R^2^ | log(likelihood)
----------: | ------: | --------: | ---------:
Zero-Inflated Negative Binomial | Complex: log(`physhealth`) | .049 | -2567.8

### Check model assumptions

Here is a plot of residuals vs. fitted values on the original `physhealth` scale.

```{r}
ggplot(sm_zinb1, aes(x = fitted, y = resid)) +
    geom_point() +
    labs(title = "Residuals vs. Fitted `physhealth`",
         subtitle = "Zero-Inflated Negative Binomial Regression model")
```

### Predictions for Harry and Sally

The predictions from this zero-inflated negative binomial regression model are obtained as follows...

```{r}
predict(mod_zinb1, newdata = hs_data, type = "response")
```

As we've seen in the past, when we use `response` as the type, the predictions fall on the original `physhealth` scale. The prediction for Harry is 6.01 days, and for Sally is 1.45 days. 

## A "hurdle" model (with Poisson)

Much of the discussion here of hurdle models comes from Clay Ford at the University of Virginia^[http://data.library.virginia.edu/getting-started-with-hurdle-models/ is an excellent introduction, by Clay Ford, a Statistical Research Consultant at the University of Virginia Library. I can also recommend https://rpubs.com/kaz_yos/pscl-2 as a place to learn more about the `pscl` package, and the fitting and interpretation of both hurdle and zero-inflated regression models. That `rpubs` site has a link to [this article by Hu, Pavlicova and Nunes from the Am J Drug Alcohol Abuse](https://www.ncbi.nlm.nih.gov/pubmed/21854279) which provides a real set of examples from a trial of a behavioral health intervention meant to reduce the risk of unprotected sexual occasions as part of a strategy to reduce HIV risk.]. Ford describes a hurdle model as follows:

> The hurdle model is a two-part model that specifies one process for zero counts and another process for positive counts. The idea is that positive counts occur once a threshold is crossed, or put another way, a hurdle is cleared. If the hurdle is not cleared, then we have a count of 0.

> The first part of the model is typically a binary logit model. This models whether an observation takes a positive count or not. The second part of the model is usually a truncated Poisson or Negative Binomial model. Truncated means we're only fitting positive counts. If we were to fit a hurdle model to our [medicare] data, the interpretation would be that one process governs whether a patient visits a doctor or not, and another process governs how many visits are made.

To fit a hurdle model, we'll use the `hurdle` function in the `pscl` package.

```{r}
mod_hur1 <- hurdle(physhealth ~ bmi_c + smoke100,
                   dist = "poisson", zero.dist = "binomial",
                   data = sm_oh_A_young)

summary(mod_hur1)
confint(mod_hur1)
```

We are using the default settings here, using the same predictors for both models:

- a **Binomial** model to predict the probability of `physhealth` = 0 given our predictors, as specified by the `zero.dist` argument in the `hurdle` function, and
- a (truncated) **Poisson** model to predict the positive-count of `physhealth` given those same predictors, as specified by the `dist` argument in the `hurdle` function.

### Comparison to a null model

To show that this model fits better than the null model (the model with intercept only), we can compare them directly with a chi-squared test. Since we have two predictors in the full model, the degrees of freedom for this test is 2.

```{r}
mod_hurnull <- hurdle(physhealth ~ 1, dist = "poisson", 
                      zero.dist = "binomial", 
                      data = sm_oh_A_young)

summary(mod_hurnull)

pchisq(2 * (logLik(mod_hur1) - logLik(mod_hurnull)), df = 2, lower.tail = FALSE)
```

### Comparison to a Poisson Model: Vuong test

```{r}
vuong(mod_hur1, mod_poiss1)
```

The hurdle model is a significant improvement over the standard Poisson model according to this test.

### Comparison to a Zero-Inflated Poisson Model: Vuong test

Is the hurdle model comparable to the zero-inflated Poisson?

```{r}
vuong(mod_hur1, mod_zip1)
```

The hurdle model is not a significant improvement over the zero-inflated Poisson model according to this test.

### The Fitted Equation

The form of the model equation for this hurdle also requires us to take two separate models into account. First we have a logistic regression model to predict the log odds of zero `physhealth` days...

```
logit(physhealth = 0) = 0.921 - 0.051 bmi_c - 0.413 smoke100
```

That takes care of the zeros. Then, to predict the number of `physhealth` days, we use the following truncated Poisson model:

```
log(physhealth) = 1.90 + 0.014 bmi_c + 0.441 smoke100
```

which is truncated to produce only estimates greater than zero.

### Interpreting the Coefficients

We can exponentiate the logistic regression coefficients to obtain results in terms of odds ratios for that model, and that can be of some help in understanding the process behind excess zeros.

Also, exponentiating the coefficients of the count model help us describe those counts on the original scale of `physhealth`.

```{r}
exp(coef(mod_hur1))
```

For example,

- in the model for `physhealth` = 0, the odds of `physhealth` = 0 are 151% as high for subjects with `smoke100` = 1 as for non-smokers with the same BMI.
- in the Poisson model for `physhealth`, the `physhealth` count is estimated to increase by 1.55 for smokers as compared to non-smokers with the same BMI. 

### Testing the Predictors

We can test the model with and without `bmi_c`, for example, by fitting the model both ways, and comparing the results with either a Wald or Likelihood Ratio test, each of which is available in the `lmtest` package.

```{r}
mod_hur1_nobmi <- hurdle(physhealth ~ smoke100,
                         dist = "poisson", 
                         zero.dist = "binomial",
                         data = sm_oh_A_young)

lmtest::waldtest(mod_hur1, mod_hur1_nobmi)
```

```{r}
lmtest::lrtest(mod_hur1, mod_hur1_nobmi)
```

### Store fitted values and residuals

The `broom` package does not work with the `hurdle` class of models. Again we need to build up the fitted values and residuals ourselves.

```{r}
sm_hur1 <- sm_oh_A_young %>%
    mutate(fitted = fitted(mod_hur1, type = "response"),
           resid = resid(mod_hur1, type = "response"))

sm_hur1 %>% 
    dplyr::select(physhealth, fitted, resid) %>%
    head()
```

### Modeled Number of Zero Counts

Once again, we can compare the observed number of zero `physhealth` results to the expected number of zero values from the likelihood-based models.

```{r}
round(c("Obs" = sum(sm_oh_A_young$physhealth == 0),
  "Poisson" = sum(dpois(0, fitted(mod_poiss1))),
  "NB" = sum(dnbinom(0, mu = fitted(mod_nb1), size = mod_nb1$theta)),
  "ZIP" = sum(predict(mod_zip1, type = "prob")[,1]),
  "ZINB" = sum(predict(mod_zinb1, type = "prob")[,1]),
  "Hurdle" = sum(predict(mod_hur1, type = "prob")[,1])),0)
```

The hurdle model does about as well as the negative binomial and zero-inflated models. All but the Poisson give reasonable fits in this regard.

### Rootogram for Hurdle Model

```{r}
countreg::rootogram(mod_hur1, max = 30)
```

The results are still not perfect, of course. In fitting the zeros exactly, we're underfitting counts of 1, 2, and 30, and overfitting many of the counts between 6 and 20. We still have a problem here with overdispersion. That's why we'll consider a hurdle model with a negative binomial regression for the counts in a moment.

### Understanding the Modeled Counts in Detail

The expected mean count uses both parts of the hurdle model. Mathematically, we want...

$$
E[y | {\bf x}] = \frac{1 - f_1(0 | {\bf x})}{1 - f_2(0 | {\bf x})} \mu_2({\bf x})
$$

where

- our count of `physhealth` is $y$
- our predictors are represented by **x**
- and the expected count is the product of a ratio and a mean.

> The ratio is the probability of a non-zero in the first process divided the probability of a non-zero in the second untruncated process. The f symbols represent distributions. Recall these are logistic and Poisson, respectively, by default but can be others. The mean is for the untruncated version of the positive-count process.

If we want to see the expected hurdle counts, we can get them using some clever applications of the `predict` function.

The first six expected mean counts ($E[y | {\bf x}]$ from the equation above) are:

```{r}
head(predict(mod_hur1, type = "response"))
```

The ratio of non-zero probabilities, $\frac{1 - f_1(0 | {\bf x})}{1 - f_2(0 | {\bf x})}$, from the mathematical expression above can be extracted by:

```{r}
head(predict(mod_hur1, type = "zero"))
```

The mean for the untruncated process, $\mu_2({\bf x})$, can also be obtained by:

```{r}
head(predict(mod_hur1, type = "count"))
```

and we can multiply these last two pieces together to verify that they match our expected hurdle counts.

```{r}
head(predict(mod_hur1, type = "zero") * predict(mod_hur1, type = "count"),5)
```

### Specify the R^2^ and log (likelihood) values

We can calculate a proxy for R^2^ as the squared correlation of the fitted values and the observed values.

```{r}
# The correlation of observed and fitted values
(hur1_r <- with(sm_hur1, cor(physhealth, fitted)))

# R-square
hur1_r^2
```

```{r}
logLik(mod_hur1)
```

Here, we have

Model | Scale | R^2^ | log(likelihood)
----------: | ------: | --------: | ---------:
Hurdle Model (Poisson) | Complex: log(`physhealth`) | .050 | -4184.1

### Check model assumptions

Here is a plot of residuals vs. fitted values on the original `physhealth` scale.

```{r}
ggplot(sm_hur1, aes(x = fitted, y = resid)) +
    geom_point() +
    labs(title = "Residuals vs. Fitted `physhealth`",
         subtitle = "Hurdle model with Poisson counts")
```

### Predictions for Harry and Sally

The predictions from this zero-inflated negative binomial regression model are obtained as follows...

```{r}
predict(mod_hur1, newdata = hs_data, type = "response")
```

As we've seen in the past, when we use `response` as the type, the predictions fall on the original `physhealth` scale. The prediction for Harry is 5.93 days, and for Sally is 1.47 days. 

## A "hurdle" model (with negative binomial for overdispersion)

Let's account for overdispersion better with a negative binomial model for the counts in our hurdle model. We specify that the positive-count process be fit with this NB model using `dist = negbin`.

```{r}
mod_hur_nb1 <- hurdle(physhealth ~ bmi_c + smoke100,
                   dist = "negbin", zero.dist = "binomial",
                   data = sm_oh_A_young)

summary(mod_hur_nb1)
confint(mod_hur_nb1)
```

### Comparison to a null model

To show that this model fits better than the null model (the model with intercept only), we can compare them directly with a chi-squared test. Since we have two predictors in the full model, the degrees of freedom for this test is 2.

```{r}
mod_hur_nb_null <- hurdle(physhealth ~ 1, dist = "negbin", 
                      zero.dist = "binomial", 
                      data = sm_oh_A_young)

summary(mod_hur_nb_null)

pchisq(2 * (logLik(mod_hur_nb1) - logLik(mod_hur_nb_null)), df = 2, lower.tail = FALSE)
```

### Comparison to a Negative Binomial Model: Vuong test

```{r}
vuong(mod_hur_nb1, mod_nb1)
```

The hurdle model is a significant improvement over the standard negative binomial model according to the raw and AIC-corrected versions of this test, but not the BIC-corrected version.

### Comparison to a Zero-Inflated NB Model: Vuong test

Is the hurdle model comparable to the zero-inflated Poisson?

```{r}
vuong(mod_hur_nb1, mod_zinb1)
```

The hurdle model is just barely a significant improvement over the zero-inflated Negative Binomial model.

### Comparing the Hurdle Models with AIC and BIC

```{r}
AIC(mod_hur1); BIC(mod_hur1)
```

```{r}
AIC(mod_hur_nb1); BIC(mod_hur_nb1)
```

The negative binomial approach certainly looks better than the Poisson here.


### The Fitted Equation

The form of the model equation for this hurdle also requires us to take two separate models into account. First we have a logistic regression model to predict the log odds of zero `physhealth` days...

```
logit(physhealth = 0) = -0.921 + 0.051 bmi_c + 0.413 smoke100
```

That takes care of the zeros. Then, to predict the number of `physhealth` days, we use the following truncated negative binomial model:

```
log(physhealth) = 1.30 + 0.018 bmi_c + 0.576 smoke100
```

which is truncated to produce only estimates greater than zero, with $\theta$ estimated as `exp(-1.123)` or `r round(exp(-1.123), 3)`.

### Interpreting the Coefficients

We can exponentiate the logistic regression coefficients to obtain results in terms of odds ratios for that model, and that can be of some help in understanding the process behind excess zeros.

```{r}
exp(coef(mod_hur_nb1))
```

For example,

- in the model for `physhealth` = 0, the odds of `physhealth` = 0 are 151% as high for subjects with `smoke100` = 1 as for non-smokers with the same BMI.

### Testing the Predictors

We can test the model with and without `bmi_c`, for example, by fitting the model both ways, and comparing the results with either a Wald or Likelihood Ratio test, each of which is available in the `lmtest` package.

```{r}
mod_hurnb1_nobmi <- hurdle(physhealth ~ smoke100,
                         dist = "negbin", 
                         zero.dist = "binomial",
                         data = sm_oh_A_young)

lmtest::waldtest(mod_hur_nb1, mod_hurnb1_nobmi)
```

```{r}
lmtest::lrtest(mod_hur_nb1, mod_hurnb1_nobmi)
```


### Store fitted values and residuals

Again we need to build up the fitted values and residuals, without `broom` to help.

```{r}
sm_hur_nb1 <- sm_oh_A_young %>%
    mutate(fitted = fitted(mod_hur_nb1, type = "response"),
           resid = resid(mod_hur_nb1, type = "response"))

sm_hur_nb1 %>% 
    dplyr::select(physhealth, fitted, resid) %>%
    head()
```

### Rootogram for NB Hurdle Model

```{r}
countreg::rootogram(mod_hur_nb1, max = 30)
```

This improves the situation, but we're still underfitting the 30s.

### Specify the R^2^ and log (likelihood) values

We can calculate a proxy for R^2^ as the squared correlation of the fitted values and the observed values.

```{r}
# The correlation of observed and fitted values
(hurnb1_r <- with(sm_hur_nb1, cor(physhealth, fitted)))

# R-square
hurnb1_r^2
```

```{r}
logLik(mod_hur_nb1)
```

Here, we have

Model | Scale | R^2^ | log(likelihood)
----------: | ------: | --------: | ---------:
Hurdle Model (Neg. Bin.) | Complex: log(`physhealth`) | .050 | -2566.3

### Check model assumptions

Here is a plot of residuals vs. fitted values on the original `physhealth` scale.

```{r}
ggplot(sm_hur_nb1, aes(x = fitted, y = resid)) +
    geom_point() +
    labs(title = "Residuals vs. Fitted `physhealth`",
         subtitle = "Hurdle model with Negative Binomial counts")
```

### Predictions for Harry and Sally

The predictions from this zero-inflated negative binomial regression model are obtained as follows...

```{r}
predict(mod_hur_nb1, newdata = hs_data, type = "response")
```

The prediction for Harry is 6.04 days, and for Sally is 1.45 days. 

### Note: Fitting a Different Hurdle Model for Counts and Pr(zero)

Suppose we wanted to use only `bmi_c` to predict the probability of a zero count, but use both predictors in the model for the positive counts. We use the `|` command.

```{r}
mod_hur_new1 <- 
    hurdle(physhealth ~ bmi_c + smoke100 | bmi_c,
           dist = "negbin", zero.dist = "binomial",
           data = sm_oh_A_young)

summary(mod_hur_new1)
```

### Hanging Rootogram for this new Hurdle Model

```{r}
countreg::rootogram(mod_hur_new1, max = 30)
```

Not a meaningful improvement, certainly.

## A Tobit (Censored) Regression Model

The idea of the **tobit** model (sometimes called a **censored regression** model) is to estimate associations for outcomes where we can see either left-censoring (censoring from below) or right-censoring (censoring from above.) 

Consider the variable `physhealth`, which is restricted to fall between 0 and 30 by the way the measure was constructed. But supposed we think about a broader and latent (unobserved) variable describing physical health. Among the people with `physhealth` = 0, some would be incredible athletes and others would be in much poorer physical health but still healthy enough to truthfully answer 0. On the other end, some of the people responding 30 are in substantially worse physical health than others with that same response. 

- Censoring from below takes place when values at or below a threshold (in this case 0) take that value.
- Censoring from above takes place when values at or above a threshold (here, 30) take that value.

Several examples of tobit analysis are available at https://stats.idre.ucla.edu/r/dae/tobit-models/, which is my primary source for the material here on those models.

The tobit model postulates that the value 0 in our model is just the lower limit of the underlying measure of poor physical health that we would actually observe in the population if we had a stronger measure. Similarly, we'll postulate that 30 is just the upper limit of "poor health" that we can see. The approach I'll take to run the tobit model comes from the `vglm` function in the `VGAM` package.

Here's the model, and its summary. Note that the default Lower value for a tobit model is 0, so we didn't technically have to list that here.

```{r}
mod_tob1 <- vglm(physhealth ~ bmi_c + smoke100, 
                 tobit(Lower = 0, Upper = 30),
                 type.fitted = "censored",
                 data = sm_oh_A_young)

summary(mod_tob1)
confint(mod_tob1)
```

### The Fitted Equation

Because we've used the censoring approach, our model will limit its predictions to the range of [0, 30], where any predictions outside that range are censored. Values below 0 are fitted as 0, and values above 30 are fitted as 30.

The model equation is 

```
physhealth = -11.00 + 0.49 bmi_c + 5.67 smoke100
```
### Interpreting the Coefficients

Tobit model regression coefficients are interpreted as we would a set of OLS coefficients, *except* that the linear effect is on the uncensored *latent variable*, rather than on the observed outcome.

In our case, 

- a one-unit increase in `bmi_c` is associated with a 0.49 day increase in the predicted value of `physhealth`, holding `smoke100` constant
- a move from `smoke100` = 0 to 1 is associated with a 5.67 day increase in the predicted value of `physhealth`, holding `bmi_c` constant
- the coefficient labeled `(Intercept):1` is the intercept for the model and is the predicted value of `physhealth` when `smoke100` = 0 and `bmi_c` = 0. Note that this value is -11, which is outside the range of `physhealth` values we observed.
- the coefficient labeled `(Intercept):2` is a statistic we can use after we exponentiate it, as follows:
    - here `(Intercept):2` = 2.82, and exp(2.82) = `r exp(2.82)`, which is analogous to the square root of the residual variance in OLS regression, which is summarized for our OLS model as `Residual standard error: 6.609`.

### Testing the Predictors

We can test the model with and without `bmi_c`, for example, by fitting the model both ways, and comparing the results with either a Wald or Likelihood Ratio test, each of which is available in the `lmtest` package.

```{r}
mod_tob_nobmi <- vglm(physhealth ~ smoke100, 
                      tobit(Lower = 0, Upper = 30),
                      type.fitted = "censored",
                      data = sm_oh_A_young)

lmtest::waldtest(mod_tob1, mod_tob_nobmi)
```

The likelihood ratio test we have used in some other settings isn't available here.

### Store fitted values and residuals

The residuals and fitted values from the tobit model can be stored and then summarized in several ways:

```{r}
sm_tob1 <- sm_oh_A_young %>%
    mutate(fitted = fitted(mod_tob1,
                           type.fitted = "censored"),
           resid = physhealth - fitted)

sm_tob1 %>% 
    dplyr::select(physhealth, fitted, resid) %>%
    head()
```

### Building Something Like a Rootogram

Building a rootogram is tricky for a tobit model, to say the least, but we can approximate a piece of the issue by plotting the rounded fitted values against the observed `physhealth` data.

```{r}
ggplot(sm_tob1, aes(x = physhealth, y = round(fitted))) +
    geom_jitter(width = 0.2) + 
    geom_abline(intercept = 0, slope = 1, col = "red")
```

Note that the model never predicts a subject to have an underlying `physhealth` worse than about 12 (remember that larger numbers are worse here.) 

### Tables of the Observed and Fitted `physhealth` from Tobit

```{r}
addmargins(table(round(sm_tob1$physhealth)))
```

```{r}
addmargins(table(round(sm_tob1$fitted)))
```

### Specify the R^2^ and log (likelihood) values

We can calculate a proxy for R^2^ as the squared correlation of the fitted values and the observed values.

```{r}
# The correlation of observed and fitted values
(tob1_r <- with(sm_tob1, cor(physhealth, fitted)))

# R-square
tob1_r^2
```

```{r}
logLik(mod_tob1)
```

Here, we have

Model | Scale | R^2^ | log(likelihood)
----------: | ------: | --------: | ---------:
Tobit | `physhealth` | .021 | -2567.3

### Check model assumptions

Here is a plot of residuals vs. fitted values.

```{r}
ggplot(sm_tob1, aes(x = fitted, y = resid)) +
    geom_point() +
    labs(title = "Residuals vs. Fitted Values for Tobit 1")
```

Here is a normal Q-Q plot of the Tobit Model 1 residuals.

```{r}
qqnorm(sm_tob1$resid)
```

### Predictions for Harry and Sally

The predictions from this tobit model are obtained as follows...

```{r}
predict(mod_tob1, newdata = hs_data, type = "response")
```

The prediction for both Harry and Sally under the tobit model would be truncated to 0 days.
