==> 23_coxmodels2.Rmd <==
# Cox Regression Models for Survival Data: Example 2

## A Second Example: The `leukem` data

```{r read in data sets you will use}
leukem
```

The data describe 51 leukemia patients. The variables are:

- `id`, a patient identification code
- `age`, age at diagnosis
- `pblasts`, the Smear differential percentage of blasts
- `pinf`, the Percentage of absolute marrow leukemia infiltrate
- `plab`, the Percentage labeling index of the bone marrow leukemia cells
- `maxtemp`, Highest temperature prior to treatment (in $^\circ F$)
- `months`, which is Survival time from diagnosis (in months)
- `alive`, which indicates Status as of the end of the study (1 = alive and thus censored, 0 = dead)

```{r glimpse of leukem}
glimpse(leukem)
```

### Creating our response: A survival time object

Regardless of how we're going to fit a survival model, we start by creating a *survival time* object that combines the information in `months` (the survival times, possibly censored) and `alive` (the censoring indicator) into a single variable we'll call `stime` in this example. 

The function below correctly registers the survival time, and censors subjects who are alive at the end of the study (we need to indicate those whose times are known, and they are identified by `alive == 0`). All other subjects are alive for at least as long as we observe them, but their exact survival times are *right-censored*.

```{r create survival object}
stime <- Surv(leukem$months, leukem$alive == 0)
stime
```

### Models We'll Fit

We'll fit several models here, including:

- Model A: A model for survival time using `age` at diagnosis alone.
- Model B: A model for survival time using the main effects of 5 predictors, specifically, `age`, `pblasts`, `pinf`, `plab`, and `maxtemp`.
- Model B2: The model we get after applying stepwise variable selection to Model B, which will include `age`, `pinf` and `plab`.
- Model C: A model using `age` (with a restricted cubic spline), `plab` and `maxtemp`

## Model A: `coxph` Model for Survival Time using `age` at diagnosis

We'll start by using `age` at diagnosis to predict our survival object (survival time, accounting for censoring).

```{r surv_leukem_modelA_fit}
modA <- coxph(Surv(months, alive==0) ~ age, 
              data=leukem, model=TRUE)

summary(modA)
```

Across these 51 subjects, we observe 45 events (deaths) and 6 subjects are censored. The hazard ratio (shown under `exp(coef)`) is `r summary(modA)$coef[2]`, and this means each additional year of age at diagnosis is associated with a 1.03-fold increase in the hazard of death. 

For this simple Cox regression model, we will focus on interpreting 

1. the **hazard ratio** (specified by the `exp(coef)` result and associated confidence interval) as a measure of effect size,
    + Here, the hazard ratio associated with a 1-year increase in `age` is `r round(summary(modA)$coef[2],3)`, and its 95\% confidence interval is: (`r round(summary(modA)$conf.int[3],3)`, `r round(summary(modA)$conf.int[4],3)`). 
    + Since this confidence interval ratio does not include 1 (barely), we can conclude that there is a significant association between `age` and `stime` at the 5\% significance level.
2. the **concordance** and **Rsquare** as measures of fit quality, and
    + **concordance** is only appropriate when we have at least one continuous predictor in our Cox model, in which case it assesses the probability of agreement between the survival time and the risk score generated by the predictor (or set of predictors.) A value of 1 indicates perfect agreement, but values of 0.6 to 0.7 are more common in survival data. 0.5 is an agreement that is no better than chance. Here, our concordance is `r round(summary(modA)$concordance[1], 3)`, which is a fairly typical value.
    + **Rsquare** in this setting is Cox and Snell's pseudo-R^2^, which reflects the improvement of the model we have fit over the model with the intercept alone - a comparison that is tested by the likelihood ratio test. The maximum value of this statistic is often less than one, in which case R will tell you that. Here, our observed pseudo-R^2^ is `r round(summary(modA)$rsq[1],3)` and that is out of a possible maximum of `r round(summary(modA)$rsq[2],3)`.
3. the significance tests, particularly the **Wald** test (shown next to the coefficient estimates in the position of a t test in linear regression), and the **Likelihood ratio** test at the bottom of the output, which compares this model to a null model which predicts the mean survival time for all subjects.
    + The Wald test for an individual predictor compares the coefficient to its standard error, just like a t test in linear regression.
    + The likelihood ratio test compares the entire model to the null model (intercept-only). Again, run an ANOVA (technically an analysis of deviance) to get more details on the likelihood-ratio test.

```{r surv_leukem_modelA_anova}
anova(modA)
```

### Plotting the Survival Curve implied by Model A

```{r surv_leukem_modelA_plot}
plot(survfit(modA), ylab="Probability of Survival",
     xlab="Months in Study", col=c("red", "black", "black"))
```

### Testing the Proportional Hazards Assumption

As we've noted, the key assumption in a Cox model is that the hazards are **proportional**. 

```{r surv_leukem_modelA_zphtest}
cox.zph(modA)
```

A significant result here would indicate a problem with the proportional hazards assumption - again, not the case here. We can also plot the results:

```{r surv_leukem_modelA_zphplot}
plot(cox.zph(modA))
```

We're looking for the smooth curve to be fairly level across the time horizon here, as opposed to substantially increasing or decreasing in level as time passes.

## Building Model A with `cph` for the `leukem` data

```{r surv_leukem_modelA_cph_fit}
units(leukem$months) <- "month"
d <- datadist(leukem)
options(datadist="d")
modA_cph <- cph(Surv(months, alive==0) ~ age, data=leukem, 
              x=TRUE, y=TRUE, surv=TRUE, time.inc=12)
```

```{r surv_leukem_modelA_cph_coefficients}
modA_cph
exp(coef(modA_cph)) # hazard ratio estimate
exp(confint(modA_cph)) # hazard ratio 95% CI
```

### Plotting the `age` effect implied by our model.

We can plot the `age` effect implied by the model, using `ggplot2`, as follows...

```{r surv_leukem_modelA_cph_ggplot}
ggplot(Predict(modA_cph, age))
```

### Survival Plots (Kaplan-Meier) of the `age` effect

The first survival plot I'll show displays 95% confidence intervals for the probability of survival at the median `age` at diagnosis in the sample, which turns out to be `r median(leukem$age)` years, with numbers of patients still at risk indicated every 12 months of time in the study. We can substitute in **conf = `bars`** to get a different flavor for this sort of plot. 

```{r surv_leukem_modelA_cph_survplot1}
survplot(modA_cph, age=median(leukem$age), conf.int=.95, 
         col='blue', time.inc=12, n.risk=TRUE,
         conf='bands', type="kaplan-meier", 
         xlab="Study Survival Time in Months")
```

Or we can generate a survival plot that shows survival probabilities over time across a range of values for `age` at diagnosis, as follows...

```{r surv_leukem_modelA_cph_survplot2}
survplot(modA_cph, levels.only=TRUE, time.inc=12, 
         type="kaplan-meier", 
         xlab="Study Survival Time in Months")
```

This plot shows a series of modeled survival probabilities, for five different diagnosis `age` levels, as identified by the labels. Generally we see that the younger the subject is at diagnosis, the longer their survival time in the study. 

### ANOVA test for the `cph`-built model for `leukem`

We can run a likelihood-ratio (drop in deviance) test of the significance of the `age` effect...

```{r surv_leukem_modelA_cph_anova}
anova(modA_cph)
```

### Summarizing the Effect Sizes from `modA_cph`

We can generate the usual summaries of effect size in this context, too.

```{r surv_leukem_modelA_cph_summaryandeffects, fig.height=3}
summary(modA_cph)
plot(summary(modA_cph))
```

As with all `rms` package effect estimates, this quantitative predictor (`age`) yields an effect comparing `age` at the 25th percentile of the sample (`age` = 35) to `age` at the 75th percentile (`age` = 61). So the hazard ratio is 2.32, with 95% CI (1.43, 3.77) for the effect of moving 26 years. Our `coxph` version of this same model showed a hazard ratio for the effect of moving just a single year.

### Validating the Cox Model Summary Statistics

```{r surv_leukem_modelA_cph_validate}
set.seed(432410); validate(modA_cph)
```

The R^2^ statistic barely moves, and neither does the Somers' d estimate, so at least in this simple model, the nominal summary statistics are likely to hold up pretty well in new data.

### Looking for Influential Points

This plot shows the influence of each point, in terms of DFBETA - the impact on the coefficient of `age` in the model were that specific point to be removed from the data set. We can also identify the row numbers of the largest (positive and negative) DFBETAs.

```{r surv_leukem_modelA_cph_influential_points}
plot(residuals(modA_cph, type="dfbeta", 
               collapse = leukem$id) ~ 
       leukem$id, main="Index Plot of DFBETA for Age", 
     type="h", ylab="DFBETA in modelA_cph")
which.max(residuals(modA_cph, type="dfbeta"))
which.min(residuals(modA_cph, type="dfbeta"))
```

The DFBETAs look very small here. Changes in the $\beta$ estimates as large as 0.002 don't have a meaningful impact in this case, so I don't see anything particularly influential.

### Checking the Proportional Hazards Assumption

As before, we can check the proportional hazards assumption with a test, or plot.

```{r surv_leukem_modelA_cph_proportional_hazards_assumption}
cox.zph(modA_cph)
plot(cox.zph(modA_cph))
```

Still no serious signs of trouble, of course. We'll see what happens when we fit a bigger model.

## Model B: Fitting a 5-Predictor Model with `coxph`

Next, we use the `coxph` function from the `survival` package to apply a Cox regression model to predict the survival time using the main effects of the five predictors: `age`, `pblasts`, `pinf`, `plab` and `maxtemp`.

```{r kitchen sink model linear}
modB <- coxph(Surv(months, alive==0) ~ 
          age + pblasts + pinf + plab + maxtemp, data=leukem)
modB
```

The Wald tests suggest that `age` and perhaps `plab` are the only terms which appear to be significant after accounting for the other predictors in the model. 

### Interpreting the Results from Model B

```{r summary of model B}
summary(modB)
```

Again, it appears that only `age` has a statistically significant effect, as last predictor in.

```{r anova of model B}
anova(modB)
```

At least taken in this order, none of the variables appear to add significant predictive value, given that we have already accounted for the preceding variables.



### Plotting the Survival Curve implied by Model B

```{r c23_modelB_survfit_plot}
plot(survfit(modB), ylab="Probability of Survival", 
     xlab="Months in Study", col=c("red", "black", "black"))
```

The crosses in the plot indicate censoring points, while the drops indicate people who have died, and are thus no longer at risk.



### Plotting the Cumulative Hazard implied by Model B

```{r c23_modelB_cumhaz_plot}
plot(survfit(modB, type="fleming"), col=c("blue", "black", "black"), 
     fun="cumhaz", ylab="Cumulative Hazard", xlab="Months in Study")
```

### Testing the Proportional Hazards Assumption

```{r c23_test_cph_modelB}
cox.zph(modB, transform="km", global=TRUE)
```

Note that we get a global test, and a separate test for each predictor. None show significant problems. We can plot the scaled Schoenfeld residuals directly with `ggcoxzph` from the `survminer` package.

```{r c23_zph_plot_modelB}
ggcoxzph(cox.zph(modB))
```

### Assessing Collinearity

Perhaps we have some collinearity here, which might imply that we could sensibly fit a smaller model, which would be appealing anyway, with only 45 actual events - we should probably be sticking to a model with no more than 2 or perhaps as many as 3 coefficients to be estimated.

```{r c23_modB_vif_check}
vif(modB)
```

The variance inflation factors don't look enormous - it may be that removing one of these variables will help make the others look more significant. Let's consider a stepwise variable selection algorithm to see what results...

- Note that the `leaps` library, which generates best subsets output, is designed for linear regression, as is the `lars` library, which generates the lasso. Either could be used here for some guidance, but not with the survival object `stime = Surv(months, age)` as the response, but instead only with `months` as the outcome, which ignores the censoring. The `step` procedure can be used on the survival object, though.

## Model B2: A Stepwise Reduction of Model B

```{r c23_step_modelB2}
step(modB)
```

The stepwise procedure lands on a model with three predictors. How does this result look, practically?

```{r c23_modelB2}
modB2 <- coxph(Surv(months, alive==0) ~ age + pinf + plab, data=leukem)
summary(modB2)
```

### The Survival Curve implied by Model B2

```{r c23_modelB2_survivalcurves}
plot(survfit(modB2), ylab="Probability of Survival", xlab="Months in Study", 
     col=c("red", "black", "black"))
```



### Checking Proportional Hazards for Model B2

```{r c23_modelB2_testzph}
cox.zph(modB2, transform="km", global=TRUE)

ggcoxzph(cox.zph(modB2))
```

## Model C: Using a Spearman Plot to pick a model

If we want to use the **Spearman** $\rho^2$ plot to consider how we might perhaps incorporate non-linear terms describing any or all of the five potential predictors (`age`, `pblasts`, `pinf`, `plab` and `maxtemp`) for survival time, we need to do so on the raw `months` variable, rather than the survival object (`stime = Surv(months, alive==0))` which accounts for censoring...

```{r c23_months_spearman2_plot}
plot(spearman2(months ~ age + pblasts + pinf + plab + maxtemp, data=leukem))
```

Recognizing that we can probably only fit a small model safely (since we observe only 45 actual [uncensored] survival times) I will consider a non-linear term in `age` (specifically a restricted cubic spline with 3 knots), along with linear terms for `plab` and `maxtemp`. I'm mostly just looking for a new model to study for this example.

### Fitting Model C

```{r c23_modelC}
# still have datadist set up for leukem
modC <- cph(Surv(months, alive==0) ~ rcs(age, 3) + plab + maxtemp, 
            data=leukem, x=TRUE, y=TRUE, surv=TRUE, time.inc=12)
modC
```

### ANOVA for Model C

```{r c23_modelC_anova}
anova(modC)
```

### Summarizing Model C Effect Sizes

```{r c23_modelC_summaries}
summary(modC)
plot(summary(modC))
```

### Plotting the diagnosis `age` effect in Model C

Of course, we're no longer assuming that the log relative hazard is linear in `age`, once we include a restricted cubic spline for `age` in our Model C. So our hazard ratio and confidence intervals for `age` are a bit trickier to understand.

```{r c23_modelC_hazardratios}
exp(coef(modC))
exp(confint(modC))
```

We can use `ggplot` and the `Predict` function to produce plots of the log Relative Hazard associated with any of our predictors, while holding the others constant at their medians. The effects of `maxtemp` and `plab` in our Model C are linear in the log Relative Hazard, but `age`, thanks to our use of a restricted cubic spline with 3 knots, shows a single bend.

```{r c23_modelC_predictions_plot, fig.height=4}
ggplot(Predict(modC, age))
```

### Survival Plot associated with Model C

Let's look at a survival plot associated with Model C for a subject with median values of our three predictors.

```{r c23_survplot_medians}
survplot(modC, age=median(leukem$age), conf.int=0.95, col="blue", 
         time.inc=12, n.risk=TRUE, conf='bands',
         xlab="Study Time in Months")
```

As before, we could fit such a plot to compare results across multiple `age` values, if desired.

### Checking the Proportional Hazards Assumption

```{r c23_modelC_zph_check, fig.height=5.5}
cox.zph(modC, transform="km", global=TRUE)

ggcoxzph(cox.zph(modC))
```

### Model C Nomogram

```{r c23_modelC_nomogram, fig.height=6}
sv <- Survival(modC)
surv12 <- function(x) sv(12, lp=x)
surv24 <- function(x) sv(24, lp=x)

plot(nomogram(modC, fun=list(surv12, surv24), 
              funlabel=c('12-month survival', '24-month survival')))
```

### Validating Model C's Summary Statistics

We can validate the model for Somers' $D_{xy}$, which is the rank correlation between the predicted log hazard and observed survival time, and for slope shrinkage.

```{r c23_modelC_validate}
set.seed(43234); validate(modC, B=100)
```

### Calibration of Model C (12-month survival estimates)

Finally, we validate the model for calibration accuracy in predicting the probability of surviving one year.

Quoting Harrell (page 529, RMS second edition):

> The bootstrap is used to estimate the optimism in how well predicted [one]-year survival from the final Cox model tracks flexible smooth estimates, without any binning of predicted survival probabilities or assuming proportional
hazards.

The `u` variable specifies the length of time at which we look at the calibration. I've specified the `units` earlier to be months.

```{r c23_modelC_calibrate, messages=FALSE, fig.height=6}
set.seed(43233); plot(rms::calibrate(modC, B = 10, u = 12))
```

The model seems neither especially well calibrated nor especially poorly so - looking at the comparison of the blue curve to the gray, our predictions basically aren't aggressive enough - more people are surviving to a year in our low predicted probability of 12 month survival group, and fewer people are surviving on the high end of the x-axis than should be the case.
